Transition to new site-specific configuration mechanism
`local-configs`, `local-sources` and `local-templates` mechanism is
replaced with new one. Libraries installed this way are no longer
used, so they are commented out, but left in the file to illustrate
the idea and re-implemnt that logic later.

There is a pseudo-hook placed now for site-specific
`site-config/site-generic.el` file to be loaded when present.

Additionally, gitignore is improved.